Construct the flow
