+++ app/src/main/res/layout/activity_connection.xml
@@ -1,0 +1,62 @@
+<?xml version="1.0" encoding="utf-8"?>
+<ScrollView
+        xmlns:android="http://schemas.android.com/apk/res/android"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent">
+    <LinearLayout
+            android:layout_width="match_parent"
+            android:layout_height="match_parent"
+            android:orientation="vertical">
+        <TextView
+                android:text="Room Code"
+                android:textSize="36sp"
+                android:layout_margin="24dp"
+                android:gravity="center_horizontal"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:id="@+id/roomCodeShow"
+                android:visibility="gone"/>
+        <Button
+                android:text="Advertise"
+                android:layout_width="match_parent"
+                android:layout_height="100dp"
+                android:id="@+id/buttonAdvertise"
+                android:layout_margin="16dp"
+                android:background="@android:color/holo_green_light"/>
+        <EditText
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:hint="Room Code"
+                android:id="@+id/roomCodeText"
+                android:layout_gravity="center_horizontal"
+                android:textSize="36sp"
+                android:layout_margin="24dp"
+                android:maxLength="4"
+                android:digits="abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789"/>
+        <Button
+                android:text="Discover"
+                android:layout_width="match_parent"
+                android:layout_height="100dp"
+                android:id="@+id/buttonDiscover"
+                android:layout_margin="16dp"
+                android:enabled="false"
+                android:background="@android:color/holo_orange_light"/>
+        <Button
+                android:text="Broadcast"
+                android:layout_width="match_parent"
+                android:layout_height="100dp"
+                android:id="@+id/buttonBroadcast"
+                android:layout_margin="16dp"
+                android:enabled="false"
+                android:visibility="gone"
+                android:background="@android:color/holo_blue_light"/>
+        <Button
+                android:text="Stop"
+                android:layout_width="match_parent"
+                android:layout_height="100dp"
+                android:id="@+id/buttonStop"
+                android:layout_margin="16dp"
+                android:enabled="false"
+                android:background="@android:color/holo_red_light"/>
+    </LinearLayout>
+</ScrollView>
