+++ app/src/main/res/layout/activity_main.xml
@@ -1,65 +1,41 @@
 <?xml version="1.0" encoding="utf-8"?>
-<ScrollView
-        xmlns:android="http://schemas.android.com/apk/res/android"
-        xmlns:tools="http://schemas.android.com/tools"
-        xmlns:app="http://schemas.android.com/apk/res-auto"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        tools:context=".MainActivity">
-    <LinearLayout
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:orientation="vertical">
-        <TextView
-                android:text="Room Code"
-                android:textSize="36sp"
-                android:layout_margin="24dp"
-                android:gravity="center_horizontal"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:id="@+id/roomCodeShow"
-                android:visibility="gone"/>
-        <Button
-                android:text="Advertise"
-                android:layout_width="match_parent"
-                android:layout_height="100dp"
-                android:id="@+id/button"
-                android:layout_margin="16dp"
-                android:background="@android:color/holo_green_light"/>
-        <EditText
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:hint="Room Code"
-                android:id="@+id/roomCodeText"
-                android:layout_gravity="center_horizontal"
-                android:textSize="36sp"
-                android:layout_margin="24dp"
-                android:maxLength="4"
-                android:digits="abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789"/>
-        <Button
-                android:text="Discover"
-                android:layout_width="match_parent"
-                android:layout_height="100dp"
-                android:id="@+id/button2"
-                android:layout_margin="16dp"
-                android:enabled="false"
-                android:background="@android:color/holo_orange_light"/>
-        <Button
-                android:text="Broadcast"
-                android:layout_width="match_parent"
-                android:layout_height="100dp"
-                android:id="@+id/button4"
-                android:layout_margin="16dp"
-                android:enabled="false"
-                android:visibility="gone"
-                android:background="@android:color/holo_blue_light"/>
-        <Button
-                android:text="Stop"
-                android:layout_width="match_parent"
-                android:layout_height="100dp"
-                android:id="@+id/button3"
-                android:layout_margin="16dp"
-                android:enabled="false"
-                android:background="@android:color/holo_red_light"/>
-    </LinearLayout>
-</ScrollView>
+<android.support.constraint.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+                                             xmlns:app="http://schemas.android.com/apk/res-auto"
+                                             xmlns:tools="http://schemas.android.com/tools"
+                                             android:layout_width="match_parent"
+                                             android:layout_height="match_parent" tools:context=".MainActivity">
+
+    <TextView
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content" android:id="@+id/txtWelcome"
+            android:text="@string/txtWelcome" android:layout_marginTop="24dp"
+            app:layout_constraintTop_toTopOf="parent" app:layout_constraintStart_toStartOf="parent"
+            android:layout_marginStart="16dp" app:layout_constraintEnd_toEndOf="parent"
+            android:layout_marginEnd="16dp" android:textSize="30sp"/>
+    <TextView
+            android:text="@string/txtName"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:id="@+id/txtEnterName"
+            app:layout_constraintTop_toBottomOf="@+id/txtWelcome" app:layout_constraintStart_toStartOf="parent"
+            android:layout_marginStart="16dp" app:layout_constraintEnd_toEndOf="parent" android:layout_marginEnd="16dp"
+            android:layout_marginTop="24dp" android:textSize="18sp"/>
+    <EditText
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:inputType="textPersonName"
+            android:text="Name"
+            android:ems="10"
+            android:id="@+id/editTextName" android:layout_marginTop="16dp"
+            app:layout_constraintTop_toBottomOf="@+id/txtEnterName" android:layout_marginStart="16dp"
+            app:layout_constraintStart_toStartOf="parent" android:layout_marginEnd="16dp"
+            app:layout_constraintEnd_toEndOf="parent"/>
+    <Button
+            android:text="@string/btnStartGame"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:id="@+id/btnStartGame" android:layout_marginTop="32dp"
+            app:layout_constraintTop_toBottomOf="@+id/editTextName" app:layout_constraintStart_toStartOf="parent"
+            android:layout_marginStart="16dp" app:layout_constraintEnd_toEndOf="parent"
+            android:layout_marginEnd="16dp" android:textSize="18sp" android:background="#03A9F4"/>
+</android.support.constraint.ConstraintLayout>
