+++ app/src/main/res/layout/activity_scoreboard.xml
@@ -15,17 +15,17 @@
     <Button
             android:text="Restart"
             android:layout_width="wrap_content"
-            android:layout_height="wrap_content" android:id="@+id/button8"
+            android:layout_height="wrap_content" android:id="@+id/buttonRestart"
             android:layout_marginTop="96dp"
             app:layout_constraintTop_toBottomOf="@+id/recyclerView" app:layout_constraintStart_toStartOf="parent"
-            android:layout_marginStart="16dp" app:layout_constraintEnd_toStartOf="@+id/button9"
+            android:layout_marginStart="16dp" app:layout_constraintEnd_toStartOf="@+id/buttonNew"
             android:layout_marginEnd="16dp"/>
     <Button
             android:text="New Game"
             android:layout_width="wrap_content"
-            android:layout_height="wrap_content" android:id="@+id/button9"
+            android:layout_height="wrap_content" android:id="@+id/buttonNew"
             android:layout_marginTop="96dp"
             app:layout_constraintTop_toBottomOf="@+id/recyclerView" app:layout_constraintEnd_toEndOf="parent"
-            android:layout_marginEnd="16dp" app:layout_constraintStart_toEndOf="@+id/button8"
+            android:layout_marginEnd="16dp" app:layout_constraintStart_toEndOf="@+id/buttonRestart"
             android:layout_marginStart="16dp"/>
 </android.support.constraint.ConstraintLayout>
