+++ app/src/main/res/layout/activty_gametemple.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+              android:layout_width="match_parent"
+              android:layout_height="match_parent">
+    <Button android:layout_width="wrap_content" android:layout_height="wrap_content"
+            android:text="Begin" android:id="@+id/buttonTest1"/>
+    <Button android:layout_width="wrap_content" android:layout_height="wrap_content"
+            android:text="End" android:id="@+id/buttonTest2"/>
+
+</LinearLayout>
